Reduce duplication with find_isinstance_check and binder.push
The resulting simplification of visit_if_expr revealed a bug in the
binder: if a frame was marked as unreachable, but then a new frame
context was begun immediately, the new frame was not marked as
unreachable, and a block visited inside the new frame would be type
checked. For now we check all frames for reachability; probably there
is a more efficient way to do this.

This commit also cleans up check_boolean_op slightly.